chore: cleanup schema modification utils
