+++ src/pages/hcp/api-docs/consul/[[...page]].tsx
@@ -8,7 +8,7 @@ import { fetchCloudApiVersionData } from 'lib/api-docs/fetch-cloud-api-version-d
 import { ApiDocsVersionData } from 'lib/api-docs/types'
 // View
 import OpenApiDocsView from 'views/open-api-docs-view'
-import { schemaModShortenHcp } from 'views/open-api-docs-view/utils/massage-schema-utilities'
+import { schemaModShortenHcp } from 'views/open-api-docs-view/utils/massage-schema-utils'
 import {
 	getStaticPaths,
 	getStaticProps as getOpenApiDocsStaticProps,
